other: move footer component to layout
